7 seconds for 10 iterations

+++ julia_meshfree_serial_basic1/interior_fluxes.jl
@@ -1,8 +1,8 @@
 
 function interior_dGx_pos(globaldata, idx, configData, max_q, min_q)
 
-    power = configData["core"]["power"]
-    limiter_flag = configData["core"]["limiter_flag"]
+    power::Float64 = configData["core"]["power"]
+    limiter_flag::Float64 = configData["core"]["limiter_flag"]
 
     sum_delx_sqr = zero(Float64)
     sum_dely_sqr = zero(Float64)
@@ -93,8 +93,8 @@ end
 
 function interior_dGx_neg(globaldata, idx, configData, max_q, min_q)
 
-    power = configData["core"]["power"]
-    limiter_flag = configData["core"]["limiter_flag"]
+    power::Float64 = configData["core"]["power"]
+    limiter_flag::Float64 = configData["core"]["limiter_flag"]
 
     sum_delx_sqr = zero(Float64)
     sum_dely_sqr = zero(Float64)
@@ -181,8 +181,8 @@ end
 
 function interior_dGy_pos(globaldata, idx, configData, max_q, min_q)
 
-    power = configData["core"]["power"]
-    limiter_flag = configData["core"]["limiter_flag"]
+    power::Float64 = configData["core"]["power"]
+    limiter_flag::Float64 = configData["core"]["limiter_flag"]
 
     sum_delx_sqr = zero(Float64)
     sum_dely_sqr = zero(Float64)
@@ -267,8 +267,8 @@ end
 
 function interior_dGy_neg(globaldata, idx, configData, max_q, min_q)
 
-    power = configData["core"]["power"]
-    limiter_flag = configData["core"]["limiter_flag"]
+    power::Float64 = configData["core"]["power"]
+    limiter_flag::Float64 = configData["core"]["limiter_flag"]
 
     sum_delx_sqr = zero(Float64)
     sum_dely_sqr = zero(Float64)
